+++ packages/bridge/bridge-react/src/router/v5.tsx
@@ -5,13 +5,13 @@ import * as ReactRouterDom from 'react-router-dom/index.js';
 import { RouterContext } from '../provider/context';
 import { LoggerInstance } from '../utils';
 
-function WraperRouter(
-  props:
-    | Parameters<typeof ReactRouterDom.BrowserRouter>[0]
-    | Parameters<typeof ReactRouterDom.MemoryRouter>[0],
-) {
+function WraperRouter(props: Record<string, unknown>) {
   const { basename, ...propsRes } = props;
   const routerContextProps = useContext(RouterContext) || {};
+  const MemoryRouter =
+    ReactRouterDom.MemoryRouter as unknown as React.ComponentType<any>;
+  const BrowserRouter =
+    ReactRouterDom.BrowserRouter as unknown as React.ComponentType<any>;
 
   LoggerInstance.debug(`WraperRouter info >>>`, {
     ...routerContextProps,
@@ -21,14 +21,14 @@ function WraperRouter(
 
   if (routerContextProps?.memoryRoute) {
     return (
-      <ReactRouterDom.MemoryRouter
+      <MemoryRouter
         {...props}
         initialEntries={[routerContextProps?.memoryRoute.entryPath]}
       />
     );
   }
   return (
-    <ReactRouterDom.BrowserRouter
+    <BrowserRouter
       {...propsRes}
       basename={routerContextProps?.basename || basename}
     />