+++ packages/enhanced/src/index.ts
@@ -12,6 +12,7 @@ export { default as FederationModulesPlugin } from './wrapper/FederationModulesP
 export { default as FederationRuntimePlugin } from './wrapper/FederationRuntimePlugin';
 export { default as AsyncBoundaryPlugin } from './wrapper/AsyncBoundaryPlugin';
 export { default as HoistContainerReferencesPlugin } from './wrapper/HoistContainerReferencesPlugin';
+export { default as TreeShakingSharedPlugin } from './wrapper/TreeShakingSharedPlugin';
 
 export const dependencies = {
   get ContainerEntryDependency() {