+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @module-federation/runtime
 
+## 0.23.0
+
+### Patch Changes
+
+- Updated dependencies [3aa3cf9]
+  - @module-federation/runtime-core@0.23.0
+  - @module-federation/sdk@0.23.0
+  - @module-federation/error-codes@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.1
+- @module-federation/error-codes@0.22.1
+- @module-federation/runtime-core@0.22.1
+
 ## 0.22.0
 
 ### Patch Changes